requirements.txt
requests==2.31.0
python-dotenv==1.0.1
instructor==1.4.2
pyjwt==2.9.0
cryptography==43.0.1
requests-toolbelt==1.0.0
pandas==2.2.3
openpyxl==3.1.5
tqdm==4.66.5
gradio==5.3.0